+++ packages/client/src/util/isOntologyObjectV2.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { OntologyObjectV2 } from "@osdk/foundry/types";
+import type { OntologyObjectV2 } from "@osdk/internal.foundry/types";
 
 export function isOntologyObjectV2(o: any): o is OntologyObjectV2 {
   return o && typeof o === "object" && typeof o.__apiName === "string"