+++ packages/legacy-client/src/client/objects/createSingleLinkStep.ts
@@ -18,7 +18,10 @@ import type { ClientContext } from "@osdk/shared.net";
 
 import type { OntologyObject, ParameterValue, SingleLink } from "../baseTypes";
 import type { GetLinkedObjectError } from "../errors";
-import { getOnlyLinkedObject } from "../net/getOnlyLinkedObject";
+import {
+  getOnlyLinkedObject,
+  getOnlyLinkedObjectNoErrors,
+} from "../net/getOnlyLinkedObject";
 import type { Result } from "../Result";
 
 export function createSingleLinkStep<T extends OntologyObject = OntologyObject>(
@@ -30,6 +33,22 @@ export function createSingleLinkStep<T extends OntologyObject = OntologyObject>(
   return {
     async get(): Promise<Result<T, GetLinkedObjectError>> {
       return getOnlyLinkedObject(
+        client,
+        sourceObjectType,
+        sourcePrimaryKey,
+        targetLinkType,
+      );
+    },
+    async fetchOneWithErrors(): Promise<Result<T, GetLinkedObjectError>> {
+      return getOnlyLinkedObject(
+        client,
+        sourceObjectType,
+        sourcePrimaryKey,
+        targetLinkType,
+      );
+    },
+    async fetchOne(): Promise<T> {
+      return getOnlyLinkedObjectNoErrors(
         client,
         sourceObjectType,
         sourcePrimaryKey,